Added more sign up steps
