+++ gmNews/LoginViewController.swift
@@ -56,16 +56,7 @@ class LoginViewController: UIViewController {
         }
     }
     @IBAction func onSignUp(_ sender: Any) {
-        var user = PFUser()
-        user.username = usernameField.text
-        user.password = passwordField.text
-        
-        user.signUpInBackground{ (success, error) in
-            if success {
-                self.performSegue(withIdentifier: "loginSegue", sender: nil)
-            } else {
-                print("Error: \(error?.localizedDescription)")
-            }
-        }
+        self.performSegue(withIdentifier: "SignInToSignUpSegue", sender: nil)
     }
+    
 }
